Move hook example to separate file
